fix infusions being region scaled (#3339)
* fix infusions being region scaled
* improve understanding calculations to happen on item stat updates

+++ plugins/paper/src/main/java/com/playmonumenta/plugins/itemstats/infusions/Quench.java
@@ -29,7 +29,7 @@ public class Quench implements Infusion {
 	}
 
 	public static double getDurationScaling(Plugin plugin, Player player) {
-		int level = plugin.mItemStatManager.getInfusionLevel(player, InfusionType.QUENCH);
+		double level = plugin.mItemStatManager.getInfusionLevel(player, InfusionType.QUENCH);
 		return 1 + DURATION_BONUS_PER_LVL * level;
 	}
 }
